backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
python-multipart>=0.0.9
moviepy>=1.0.3
Pillow>=10.4.0
